Added exception handling

+++ src/main/java/com/zone24x7/projects/FixedProject.java
@@ -10,7 +10,7 @@ public class FixedProject extends Project {
 
 
     @Override
-    public Integer getCost() {
+    public double getCost() {
         return cost;
     }
