+++ src/main/java/com/zone24x7/projects/VariableProject.java
@@ -3,7 +3,7 @@ package com.zone24x7.projects;
 public class VariableProject extends  Project{
 
     @Override
-    public Integer getCost() {
+    public double getCost() {
         return this.getDuration()*this.getNumber_of_emp();
     }
